Fix remaining gcc-14 warnings: P1 assert sign-compare, boost array-bounds
- Cast Eigen size() in the dofs/global assert in MPI P1.h.
- Suppress GCC's false-positive -Warray-bounds around the Boost.Container
  includes in Types.h (diagnostic is attributed to the Boost headers).

+++ src/Rodin/MPI/Variational/P1/P1.h
@@ -575,7 +575,7 @@ namespace Rodin::Variational
               continue;
 
             const auto& dofs = m_fes.getDOFs(0, *i);
-            assert(dofs.size() == global.size());
+            assert(static_cast<size_t>(dofs.size()) == global.size());
 
             for (size_t k = 0; k < static_cast<size_t>(global.size()); ++k)
             {
